+++ versionmanager/semantic/parser/flat/flatparser.go
@@ -26,7 +26,7 @@ import (
 
 	"github.com/tofuutils/tenv/config"
 	"github.com/tofuutils/tenv/pkg/loghelper"
-	"github.com/tofuutils/tenv/versionmanager/semantic/parser/types"
+	"github.com/tofuutils/tenv/versionmanager/semantic/types"
 )
 
 func RetrieveVersion(filePath string, conf *config.Config) (string, error) {
