+++ versionmanager/semantic/parser/toml/tomlparser.go
@@ -26,7 +26,7 @@ import (
 	"github.com/BurntSushi/toml"
 	"github.com/tofuutils/tenv/config"
 	"github.com/tofuutils/tenv/pkg/loghelper"
-	"github.com/tofuutils/tenv/versionmanager/semantic/parser/types"
+	"github.com/tofuutils/tenv/versionmanager/semantic/types"
 )
 
 const versionName = "version"
